Fixes fitness bug that was crashing for Parajke stat on dDEM obs.
Parajke processing assumes 3 dimensions in obs data, which works
fine for MODIS (time, rows, cols) but was causing an error in the
call to sub2ind due to incorrect indexing when dDEM obs (rows,
cols) were being compared.  In this case, have temporarily forced
a degenerate 3rd dimension of 1 (time step) so that indexing is
done correctly.

Note that call to squeeze for mod and obs at end will squeeze all
dimensions of size 1, so if there is ever a case where a vector
is input, the resulting obs and mod arrays may be squeezed too far.